vcs_info: choose backend by basedir
Previously, vcs_info iterated over enabled VCS backends and output repository
status from the first backend that works (i.e. first backend that can find a
repository).

But I prefer a different behavior: I want to see the status of the repository
closest to the current working directory. For example: if there is a Mercurial
repository inside a Git repository, and I 'cd' into the Mercurial repository,
I want to see the status of Mercurial repo, even if 'git' comes before 'hg' in
the 'enable' list.

This patch adds a new algorithm for vcs_info to choose the backend:

1. Among backends whose basedir isn't an ancestor of the current directory,
old algorithm is used: whichever comes first in 'enable' list.

2. If all backends return basedirs that are ancestors of the current directory,
the one with basedir closest to the current directory is chosen.

As far as I know, unless the user has set basedir manually (GIT_WORK_TREE),
all backend return basedirs that are ancestors of the current directory.
Backends that return non-ancestor basedirs have higher priority to show the
status of that manually-configured repository.

The algorithm used by vcs_info can be configured using 'choose-closest-backend'
zstyle:

    zstyle ':vcs_info:*' choose-closest-backend yes

By default, old algorithm is used.

It can also be configured per-backend:

    zstyle ':vcs_info:hg:*' choose-closest-backend no

This effectively moves the backend into group (1), as if its basedir is never
an ancestor of the current directory. Not sure how useful this is though.

Signed-off-by: Aleksandr Mezin <mezin.alexander@gmail.com>